Created delivery generator. Made delivery just a simple value object
